+++ TruongHoc-main/BackEnd/database/migrations/2023_01_01_000007_create_lopsinhhoat_table.php
@@ -8,16 +8,18 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('lopsinhhoat', function (Blueprint $table) {
-            $table->increments('LopSHID');
-            $table->string('TenLop', 50)->unique();
-            $table->unsignedInteger('KhoaID');
-            $table->unsignedInteger('GiangVienID')->nullable(); // Trưởng khoa hoặc GVCN
-            $table->timestamps();
+        if (!Schema::hasTable('lopsinhhoat')) {
+            Schema::create('lopsinhhoat', function (Blueprint $table) {
+                $table->increments('LopSHID');
+                $table->string('TenLop', 50)->unique();
+                $table->unsignedInteger('KhoaID');
+                $table->unsignedInteger('GiangVienID')->nullable(); // Trưởng khoa hoặc GVCN
+                $table->timestamps();
 
-            $table->foreign('KhoaID')->references('KhoaID')->on('khoa')->onDelete('cascade');
-            $table->foreign('GiangVienID')->references('GiangVienID')->on('giangvien')->onDelete('set null'); // Đã bỏ comment
-        });
+                $table->foreign('KhoaID')->references('KhoaID')->on('khoa')->onDelete('cascade');
+                $table->foreign('GiangVienID')->references('GiangVienID')->on('giangvien')->onDelete('set null'); // Đã bỏ comment
+            });
+        }
     }
     public function down(): void
     {
