+++ TruongHoc-main/BackEnd/database/migrations/2023_01_01_000020_create_admin_table.php
@@ -8,15 +8,17 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('admin', function (Blueprint $table) {
-            $table->increments('AdminID');
-            $table->unsignedInteger('UserID')->nullable();
-            $table->string('HoTen', 100);
-            $table->string('Email')->unique()->nullable();
-            $table->timestamps();
+        if (!Schema::hasTable('admin')) {
+            Schema::create('admin', function (Blueprint $table) {
+                $table->increments('AdminID');
+                $table->unsignedInteger('UserID')->nullable();
+                $table->string('HoTen', 100);
+                $table->string('Email')->unique()->nullable();
+                $table->timestamps();
 
-            $table->foreign('UserID')->references('UserID')->on('users')->onDelete('set null');
-        });
+                $table->foreign('UserID')->references('UserID')->on('users')->onDelete('set null');
+            });
+        }
     }
     public function down(): void
     {
